+++ lib/dense/path.rb
@@ -160,14 +160,17 @@ class Dense::Path
 
     attr_reader :container_class, :root_path, :remaining_path
 
-    def initialize(container, root_path, remaining_path)
+    def initialize(container, root_path, remaining_path, message=nil)
 
       @container_class = container.is_a?(Class) ? container : container.class
 
       @root_path = Dense::Path.make(root_path)
       @remaining_path = Dense::Path.make(remaining_path)
 
-      if @root_path
+      if message
+        super(
+          message)
+      elsif @root_path
         super(
           "Found nothing at #{fail_path.to_s.inspect} " +
           "(#{@remaining_path.original.inspect} remains)")
@@ -180,7 +183,15 @@ class Dense::Path
 
     def expand(root_path)
 
-      err = self.class.new(container_class, root_path, remaining_path)
+      err = self.class.new(container_class, root_path, remaining_path, nil)
+      err.set_backtrace(self.backtrace)
+
+      err
+    end
+
+    def relabel(message)
+
+      err = self.class.new(container_class, root_path, remaining_path, message)
       err.set_backtrace(self.backtrace)
 
       err